refactor: apply new gov/params structure (#4)
- Remove `x/params` dependency
- Remove legacy proposal handlers since we're not using `x/gov`
- `x/stakeibc` had `AddValidatorsProposal` and `ToggleLSMProposal`
before and there's no need to keep these two anymore
- Applied new module parameters semantics

<!--
*** Please remove the following help text before submitting: ***

Pull requests without a rationale and clear improvement may be closed
immediately.

-->

<!--
Please provide clear motivation for your patch and explain how it
improves
initia user experience or initia developer experience
significantly:

* Any test improvements or new tests that improve coverage are always
welcome.
* All other changes should have accompanying unit tests (see
`src/test/`) or
functional tests (see `test/`). Contributors should note which tests
cover
modified code. If no tests exist for a region of modified code, new
tests
  should accompany the change.
* Bug fixes are most welcome when they come with steps to reproduce or
an
explanation of the potential issue as well as reasoning for the way the
bug
  was fixed.
* Features are welcome, but might be rejected due to design or scope
issues.
If a feature is based on a lot of dependencies, contributors should
first
  consider building the system outside of initia, if possible.
-->

<!--
Initia has a thorough review process and even the most trivial change
needs to pass a lot of eyes and requires non-zero or even substantial
time
effort to review. There is a huge lack of active reviewers on the
project, so
patches often sit for a long time.
-->